feat: remove optional tcp in websocket connection
