+++ packages/timeline-state-resolver-types/src/index.ts
@@ -23,7 +23,7 @@ import { TimelineContentSingularLiveAny } from './integrations/singularLive'
 import { TimelineContentVMixAny } from './integrations/vmix'
 import { TimelineContentOBSAny } from './integrations/obs'
 import { TimelineContentTriCasterAny } from './integrations/tricaster'
-import { TimelineContentWebSocketTcpClientAny } from './integrations/websocketTcpClient'
+import { TimelineContentWebSocketClientAny } from './integrations/websocketClient'
 
 export * from './integrations/abstract'
 export * from './integrations/atem'
@@ -47,7 +47,7 @@ export * from './integrations/tricaster'
 export * from './integrations/telemetrics'
 export * from './integrations/multiOsc'
 export * from './integrations/viscaOverIP'
-export * from './integrations/websocketTcpClient'
+export * from './integrations/websocketClient'
 
 export * from './device'
 export * from './mapping'
@@ -90,7 +90,7 @@ export enum DeviceType {
 	TRICASTER = 'TRICASTER',
 	MULTI_OSC = 'MULTI_OSC',
 	VISCA_OVER_IP = 'VISCA_OVER_IP',
-	WEBSOCKET_TCP_CLIENT = 'WEBSOCKET_TCP_CLIENT',
+	WEBSOCKET_CLIENT = 'WEBSOCKET_CLIENT',
 }
 
 export interface TSRTimelineKeyframe<TContent> extends Omit<Timeline.TimelineKeyframe, 'content'> {
@@ -152,7 +152,7 @@ export type TSRTimelineContent =
 	| TimelineContentVIZMSEAny
 	| TimelineContentTelemetricsAny
 	| TimelineContentTriCasterAny
-	| TimelineContentWebSocketTcpClientAny
+	| TimelineContentWebSocketClientAny
 
 /**
  * A simple key value store that can be referred to from the timeline objects
