+++ packages/timeline-state-resolver/src/manifest.ts
@@ -56,8 +56,8 @@ import VizMSEMappings = require('./$schemas/generated/vizMSE/mappings.json')
 import VMixOptions = require('./$schemas/generated/vmix/options.json')
 import VMixMappings = require('./$schemas/generated/vmix/mappings.json')
 import VMixActions = require('./$schemas/generated/vmix/actions.json')
-import WebSocketTCPClientOptions = require('./$schemas/generated/websocketTcpClient/options.json')
-import WebSocketTCPClientActions = require('./$schemas/generated/websocketTcpClient/actions.json')
+import WebSocketClientOptions = require('./$schemas/generated/websocketClient/options.json')
+import WebSocketClientActions = require('./$schemas/generated/websocketClient/actions.json')
 
 import CommonOptions = require('./$schemas/common-options.json')
 import { generateTranslation } from './lib'
@@ -216,10 +216,10 @@ export const manifest: TSRManifest = {
 			configSchema: JSON.stringify(VMixOptions),
 			mappingsSchemas: stringifyMappingSchema(VMixMappings),
 		},
-		[DeviceType.WEBSOCKET_TCP_CLIENT]: {
-			displayName: generateTranslation('Websocket+TCP Client'),
-			actions: WebSocketTCPClientActions.actions.map(stringifyActionSchema),
-			configSchema: JSON.stringify(WebSocketTCPClientOptions),
+		[DeviceType.WEBSOCKET_CLIENT]: {
+			displayName: generateTranslation('Websocket Client'),
+			actions: WebSocketClientActions.actions.map(stringifyActionSchema),
+			configSchema: JSON.stringify(WebSocketClientOptions),
 			mappingsSchemas: {},
 		}
 	},
