+++ packages/timeline-state-resolver/src/service/devices.ts
@@ -19,7 +19,7 @@ import { TelemetricsDevice } from '../integrations/telemetrics'
 import { TriCasterDevice } from '../integrations/tricaster'
 import { SingularLiveDevice } from '../integrations/singularLive'
 import { MultiOSCMessageDevice } from '../integrations/multiOsc'
-import { WebSocketTcpClientDevice } from '../integrations/websocketTcpClient'
+import { WebSocketClientDevice } from '../integrations/websocketClient'
 
 export interface DeviceEntry {
 	deviceClass: new (context: DeviceContextAPI<any>) => Device<any, any, any>
@@ -48,7 +48,7 @@ export type ImplementedServiceDeviceTypes =
 	| DeviceType.TRICASTER
 	| DeviceType.QUANTEL
 	| DeviceType.VISCA_OVER_IP
-	| DeviceType.WEBSOCKET_TCP_CLIENT
+	| DeviceType.WEBSOCKET_CLIENT
 
 // TODO - move all device implementations here and remove the old Device classes
 export const DevicesDict: Record<ImplementedServiceDeviceTypes, DeviceEntry> = {
@@ -166,10 +166,10 @@ export const DevicesDict: Record<ImplementedServiceDeviceTypes, DeviceEntry> = {
 		deviceName: (deviceId: string) => 'VISCAOverIP ' + deviceId,
 		executionMode: () => 'sequential',
 	},
-	[DeviceType.WEBSOCKET_TCP_CLIENT]: {
-		deviceClass: WebSocketTcpClientDevice,
+	[DeviceType.WEBSOCKET_CLIENT]: {
+		deviceClass: WebSocketClientDevice,
 		canConnect: true,
-		deviceName: (deviceId: string) => 'WebSocket TCP Client ' + deviceId,
+		deviceName: (deviceId: string) => 'WebSocket Client ' + deviceId,
 		executionMode: () => 'sequential',
 	}
 }
